[REFACTOR/29] 코드리뷰 반영 weight -> padding으로 변경 및 조절 - Column -> Box로 변경 - width, height 강제 지정 삭제
